Move tableNeuronsFromNeuronInfos from neuron.utils.ts to neurons-table.utils.ts
