+++ frontend/src/tests/lib/utils/neuron.utils.spec.ts
@@ -63,7 +63,6 @@ import {
   neuronVotingPower,
   neuronsVotingPower,
   sortNeuronsByStake,
-  tableNeuronsFromNeuronInfos,
   topicsToFollow,
   userAuthorizedNeuron,
   validTopUpAmount,
@@ -98,7 +97,7 @@ import {
   type ProposalInfo,
   type RewardEvent,
 } from "@dfinity/nns";
-import { ICPToken, TokenAmount, TokenAmountV2 } from "@dfinity/utils";
+import { ICPToken, TokenAmount } from "@dfinity/utils";
 import { get } from "svelte/store";
 
 describe("neuron-utils", () => {
@@ -2832,88 +2831,6 @@ describe("neuron-utils", () => {
       expect(getTopicSubtitle({ topic: Topic.SubnetRental, i18n: en })).toBe(
         "All proposals related to renting a subnet, for example a subnet rental request."
       );
-    });
-  });
-
-  describe("tableNeuronsFromNeuronInfos", () => {
-    it("should convert neuronInfos to tableNeurons", () => {
-      const neuronId1 = 42n;
-      const neuronId2 = 342n;
-      const stake1 = 500_000_000n;
-      const stake2 = 600_000_000n;
-      const dissolveDelay1 = 15778800n;
-      const dissolveDelay2 = 252460800n;
-      const neuronInfo1 = {
-        ...mockNeuron,
-        neuronId: neuronId1,
-        fullNeuron: {
-          ...mockNeuron.fullNeuron,
-          cachedNeuronStake: stake1,
-        },
-        dissolveDelaySeconds: dissolveDelay1,
-      };
-      const neuronInfo2 = {
-        ...mockNeuron,
-        neuronId: neuronId2,
-        fullNeuron: {
-          ...mockNeuron.fullNeuron,
-          cachedNeuronStake: stake2,
-        },
-        dissolveDelaySeconds: dissolveDelay2,
-      };
-      const neuronInfos = [neuronInfo1, neuronInfo2];
-      const tableNeurons = tableNeuronsFromNeuronInfos(neuronInfos);
-      expect(tableNeurons).toEqual([
-        {
-          rowHref: "/neuron/?u=qhbym-qaaaa-aaaaa-aaafq-cai&neuron=42",
-          domKey: "42",
-          neuronId: "42",
-          stake: TokenAmountV2.fromUlps({
-            amount: 500_000_000n,
-            token: ICPToken,
-          }),
-          dissolveDelaySeconds: dissolveDelay1,
-        },
-        {
-          rowHref: "/neuron/?u=qhbym-qaaaa-aaaaa-aaafq-cai&neuron=342",
-          domKey: "342",
-          neuronId: "342",
-          stake: TokenAmountV2.fromUlps({
-            amount: 600_000_000n,
-            token: ICPToken,
-          }),
-          dissolveDelaySeconds: dissolveDelay2,
-        },
-      ]);
-    });
-
-    it("should convert neuronInfo for spawning neuron without href", () => {
-      const neuronId = 52n;
-      const dissolveDelaySeconds = BigInt(5 * SECONDS_IN_DAY);
-      const spawningNeuronInfo = {
-        ...mockNeuron,
-        neuronId: neuronId,
-        state: NeuronState.Spawning,
-        fullNeuron: {
-          ...mockNeuron.fullNeuron,
-          cachedNeuronStake: 0n,
-          spawnAtTimesSeconds: 12_312_313n,
-        },
-        dissolveDelaySeconds,
-      };
-      const neuronInfos = [spawningNeuronInfo];
-      const tableNeurons = tableNeuronsFromNeuronInfos(neuronInfos);
-      expect(tableNeurons).toEqual([
-        {
-          domKey: "52",
-          neuronId: "52",
-          stake: TokenAmountV2.fromUlps({
-            amount: 0n,
-            token: ICPToken,
-          }),
-          dissolveDelaySeconds,
-        },
-      ]);
     });
   });
 });
